Add address information to message_request_join
The address information is required to map the source
network identifier to the local port. This is not very
elegant but due to our asynchronous design we do not
actually know on which port a particular message was
received. We can use getpeername() to match the information
with the available ports.