Change cond_var_timedwait() signature
This commit changes the signature of cond_var_timedwait()
such that the last arugment is a struct timespec pointer.
Since pthread_cond_timedwait() takes the absolute time as
the last argument it is better to let users pass the absolute
time to cond_var_timedwait() as well instead of a relative
timeout value.