Bug fixes for signal handling
In order to ensure that the MESSAGE_TYPE_EXIT messages arrive at
all processes we need to wait sufficiently long (at least at the
moment where we have no proper implementation of comm_flush()).
Moreover it is not sufficient to just wait for the sendq to become
empty since we need to make sure that we are not leaving partially
routed packages in the pipeline when terminating.