(Try to) read backlog from config
In the context of the master process we can read the backlog value
from the configuration file. Since spawn->opts is NULL on recently
spawned processes we however cannot easily control this value.
The same applies to the buffer pool size and the sizes of the send
and receive queues in struct comm.