Enable spawn to run across hosts
The previous code versions only worked on a single host
since the listening socket was always bound to the loopback
interface. In order to support a network setup we now
dynamically open sockets as needed to connected to the
hosts in the tree. We also support multiple listening
sockets to handle complicated network setups. In order to
have the options available in _main_on_others() we moved
the join operation out of loop().