Provide tasks with send()/recv() functionality
This commit adds two new API functions

    task_plugin_api_send()
    task_plugin_api_recv()

that allow task plugins to send and reveive user messages. The
main loop is adapted to call task_enqueue_message() when receiving
user messages to forward them to the task plugin.
A big drawback of the current implementation is that task plugins need
to poll in order to wait for messages.