Bug fix for the exec and pmiexec plugin
The previous implementation lead to lost output if
the application wrote more than 2*MAX_LINE_LEN bytes
in one row. This is fixed by this commit. We still
loose output if the I/O rates is higher than the
processing rate of the communication thread. This is
a fundamental design flaw in the current implementation
caused by the limited capacity of the communication
queues and improper handling of the resulting error
(we really should wait in this case and not give up
right away).